Add files via upload
*DISCLAIMER* I'm a musicology student and this is the first time that i'm doing anything with coding. 
This graph shows that although Film Scores and Classical Music share a lot of the same instrumentation (ochestra), they do not necessarily share the same energy level. This is an interesting first insight into my corpus of comparing film music and classical music and the variables in which they are similar or different according to spotify.